"Confirm Show" rename to "YesNo"
